requirements.txt
matrix-nio[encryption]==0.24.0
pymongo==4.6.1
pydantic==2.5.0
click==8.1.7
python-dotenv==1.0.0
